continued general logic work all classes

+++ BJ/Player.cs
@@ -3,11 +3,11 @@
   public class Player
   {
     public string? Name { get; set; }
-    public int CurrentMoney { get; set; } = 0;
+    public decimal CurrentMoney { get; set; } = 0;
     public List<Card> Cards { get; set; } = new List<Card>();
     public int HandValue => Cards.Sum(c => c.CardValue);
     public bool InHand { get; set; } = true;
-    public int PreviousBet { get; set; }
+    public decimal PreviousBet { get; set; }
 
     public Player(string name)
     {
